refactor: move stream helpers to a different file
